+++ AADWebApp/Interfaces/IPrescriptionService.cs
@@ -10,8 +10,8 @@ namespace AADWebApp.Interfaces
     {
         public IEnumerable<Prescription> GetPrescriptions(short? id = null);
         public IEnumerable<Prescription> GetPrescriptionsByPatientId(string id = null);
-        public int CreatePrescription(int medicationId, string patientId, int dosage, DateTime dateStart, DateTime dateEnd, IssueFrequency issueFrequency);
-        public Task<int> CancelPrescriptionAsync(int id);
-        public int SetPrescriptionStatus(int id, PrescriptionStatus prescriptionStatus);
+        public int CreatePrescription(short medicationId, string patientId, int dosage, DateTime dateStart, DateTime dateEnd, IssueFrequency issueFrequency);
+        public Task<int> CancelPrescriptionAsync(short id);
+        public int SetPrescriptionStatus(short id, PrescriptionStatus prescriptionStatus);
     }
 }
