+++ AADWebApp/Models/BloodTestRequest.cs
@@ -5,9 +5,9 @@ namespace AADWebApp.Models
 {
     public class BloodTestRequest
     {
-        public int Id { get; set; }
-        public int PrescriptionId { get; set; }
-        public int BloodTestId { get; set; }
+        public short Id { get; set; }
+        public short PrescriptionId { get; set; }
+        public short BloodTestId { get; set; }
         public DateTime AppointmentTime { get; set; }
         public BloodTestRequestStatus BloodTestStatus { get; set; }
     }
